requirements.txt
cramjam==2.8.3
fastparquet==2024.2.0
fsspec==2024.3.1
kafka-python==2.0.2
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pyarrow==16.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
